backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0

# External APIs
amadeus==12.0.0

# ML and Forecasting
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.5
tensorflow==2.15.0
xgboost==2.0.2
lightgbm==4.1.0

# Data Processing
scipy==1.11.4
joblib==1.3.2

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
